[Update]: build a spring boot application
