+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "server/opensourcecobol4j"]
-	path = server/opensourcecobol4j
+	path = server/app/opensourcecobol4j
 	url = https://github.com/yutaro-sakamoto/opensourcecobol4j
