+++ server/app/.gitignore
@@ -1,0 +1,2 @@
+src/main/java/cobol4j/aws/web/sample*.java
+lib/
