+++ server/app/src/main/java/cobol4j/aws/web/App.java
@@ -3,12 +3,12 @@
  */
 package cobol4j.aws.web;
 
-public class App {
-    public String getGreeting() {
-        return "Hello World!";
-    }
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+@SpringBootApplication
+public class App {
     public static void main(String[] args) {
-        System.out.println(new App().getGreeting());
+        SpringApplication.run(App.class, args);
     }
 }
